add julia fix
